+++ data/public-did:web:digitalcredentials.github.io.json
@@ -1,0 +1,15 @@
+{
+  "@context": [
+    "https://www.w3.org/ns/did/v1",
+    "https://w3id.org/security/suites/ed25519-2020/v1"
+  ],
+  "id": "did:web:digitalcredentials.github.io",
+  "assertionMethod": [
+    {
+      "id": "did:web:digitalcredentials.github.io#z6MkrXSQTybtqyMasfSxeRBJxDvDUGqb7mt9fFVXkVn6xTG7",
+      "type": "Ed25519VerificationKey2020",
+      "controller": "did:web:digitalcredentials.github.io",
+      "publicKeyMultibase": "zD5BMsjMTWRs7mAcFxrDU78NDehZjhtdnyEabvDp63EUj"
+    }
+  ]
+}
