+++ src/issuer.ts
@@ -1,54 +1,22 @@
-import { DIDDocument } from "./types"
+import { DIDDocument } from "./types";
+import { getCustomLoaderProto } from "./common";
 import { SignatureOptions, getSigningKeyIdentifier, getSigningDate, getProofProperty } from "./signatures";
 import { default as demoCredential } from "./demoCredential.json";
 import { v4 as uuidv4 } from 'uuid';
-import { contexts as ldContexts, documentLoaderFactory } from '@transmute/jsonld-document-loader';
 import { Ed25519VerificationKey2020 } from '@digitalbazaar/ed25519-verification-key-2020';
 import { Ed25519Signature2020 } from '@digitalbazaar/ed25519-signature-2020';
-import DccContextV1 from "./contexts/dcc-v1.json";
 
-const didContext = require('did-context');
-const ed25519 = require('ed25519-signature-2020-context');
 const vc = require('@digitalbazaar/vc');
-const didKey = require('@digitalbazaar/did-method-key');
-const proofPurposes = require('jsonld-signatures').purposes;
-const DccContextV1Url = "https://w3id.org/dcc/v1";
 const VerificationMethod = "verificationMethod";
 const Challenge = "challenge";
 
-export function getController(fullDid: string) {
-  return fullDid.split('#')[0];
-}
+export function create(unlockedDID: DIDDocument) {
 
-export function createIssuer(unlockedDID: DIDDocument) {
-
-  const customLoaderProto = documentLoaderFactory.pluginFactory
-    .build({
-      contexts: {
-        ...ldContexts.W3C_Verifiable_Credentials,
-        ...ldContexts.W3ID_Security_Vocabulary,
-        ...ldContexts.W3C_Decentralized_Identifiers
-      },
-    })
-    .addContext({ [ed25519.constants.CONTEXT_URL]: ed25519.contexts.get(ed25519.constants.CONTEXT_URL) })
-    .addContext({ [didContext.constants.DID_CONTEXT_URL]: didContext.contexts.get(didContext.constants.DID_CONTEXT_URL) })
-    .addContext({ [DccContextV1Url]: DccContextV1 });
-
-  const didKeyDriver = didKey.driver();
-
-  const customLoader = customLoaderProto.
-    addResolver({
+  const customLoader = getCustomLoaderProto()
+    .addResolver({
       [unlockedDID.id]: {
         resolve: async (_did: string) => {
           return unlockedDID;
-        },
-      },
-    })
-    .addResolver({
-      ['did:key:']: {
-        resolve: async (_did: string) => {
-          return didKeyDriver.get({ did: _did });
-
         },
       },
     })
@@ -58,37 +26,13 @@ export function createIssuer(unlockedDID: DIDDocument) {
     [unlockedDID.assertionMethod[0].id, new Ed25519VerificationKey2020(unlockedDID.assertionMethod[0])]
   ]);
 
-  async function resolveDid(assertionMethod: string): Promise<any> {
-    if (assertionMethod == null) {
-      throw new Error(`Can't resolve empty DID string`);
-    }
-    const controller = getController(assertionMethod);
-    const doc = await customLoader(controller);
-    console.log(JSON.stringify(doc, null, 2));
-    return doc.document;
-    /*
-    if (!controller.startsWith('did:key')) {
-      throw new Error(`DID method (${controller}) not supported`);
-    } 
-
-    const res = await didKeyDriver.get({ did: controller });
-    if (res == null) {
-      throw new Error(`Unable to resolve DID method for ${controller}`);
-    }
-    return res;*/
-
-  }
   async function createKey(assertionMethod: string): Promise<Ed25519VerificationKey2020> {
     const keyInfo: Ed25519VerificationKey2020 | undefined = unlockedAssertionMethods.get(assertionMethod);
     if (keyInfo == null) {
-      const res = await resolveDid(assertionMethod);
-      const key = res.verificationMethod[0];
-      return new Ed25519VerificationKey2020(key);
+      throw new Error(`key for assertionMethod ${assertionMethod} not found`);
     }
-
-    return keyInfo;
+    return new Ed25519VerificationKey2020(keyInfo);
   }
-
 
   async function createSigningKey(options: SignatureOptions): Promise<Ed25519Signature2020> {
     const signingKey = await createKey(getSigningKeyIdentifier(options));
@@ -97,45 +41,6 @@ export function createIssuer(unlockedDID: DIDDocument) {
       date: getSigningDate(options)
     });
     return signatureSuite;
-  }
-
-  async function createVerificationKey(options: SignatureOptions): Promise<Ed25519VerificationKey2020> {
-    const signingKey = await createKey(getSigningKeyIdentifier(options));
-    return signingKey;
-  }
-
-  const assertionController = {
-    '@context': 'https://w3id.org/security/v2',
-    id: 'https://example.edu/issuers/565049',
-    // actual keys are going to be added in the test suite before() block
-    assertionMethod: [],
-    authentication: []
-  };
-  module.exports = assertionController;
-
-  async function verify(verifiableCredential: any, options: SignatureOptions): Promise<any> {
-    const verificationMethod = getSigningKeyIdentifier(options);
-    const didDocument = await resolveDid(verificationMethod);
-
-    const key = didDocument.verificationMethod ? didDocument.verificationMethod[0] : didDocument.assertionMethod[0];
-    const verificationKey = new Ed25519VerificationKey2020(key);
-    try {
-      /*
-      let valid = await vc.verifyCredential({
-        credential: verifiableCredential,
-        controller: didDocument,
-        suite: verificationKey,
-        customLoader
-      });
-      return valid;*/
-
-      const valid = await vc.verifyCredential({ credential: verifiableCredential, suite: verificationKey });
-      return valid;
-    }
-    catch (e) {
-      console.error(e);
-      throw e;
-    }
   }
 
   async function sign(credential: any, options: SignatureOptions): Promise<any> {
@@ -184,20 +89,10 @@ export function createIssuer(unlockedDID: DIDDocument) {
     return result;
   }
 
-  async function verifyPresentation(verifiablePresentation: any, options: SignatureOptions): Promise<any> {
-    const suite = await createVerificationKey(options);
-
-    let valid = await vc.verify({
-      presentation: { ...verifiablePresentation },
-      documentLoader: customLoader,
-      suite: suite,
-      challenge: options.challenge!,
-    });
-    return valid;
-  }
 
   async function requestDemoCredential(verifiablePresentation: any, skipVerification = false): Promise<any> {
 
+    /*
     if (!skipVerification) {
       // issuer also needs to check if challenge is expected
       const verificationOptions = {
@@ -218,18 +113,17 @@ export function createIssuer(unlockedDID: DIDDocument) {
     copy.id = uuidv4();
     copy.credentialSubject.id = subjectDid;
     copy.issuanceDate = new Date().toISOString();
-    return sign(copy, options);
+    return sign(copy, options);*/
+    return null;
+    // TODO
   }
 
   return {
     createKey,
     createSuite: createSigningKey,
-    verify,
     sign,
     signPresentation,
     createAndSignPresentation,
-    verifyPresentation,
-    requestDemoCredential,
-    customLoader // tODO
+    requestDemoCredential
   }
 }
